requirements.txt
kafka-python==2.3.0
pandas==2.3.3
requests==2.32.5
python-dotenv==1.2.1
pyyaml==6.0.3